requirements.txt
# Data manipulation and analysis
pandas
matplotlib
seaborn
plotly
geopandas
shapely
nbformat

# Jupyter
jupyter
ipython

# Agentic - Core frameworks (Langchain)
langchain
langchain-google
langchain-community
langchain-google-genai
langchain-openai
langgraph

# Agentic - supporting dependencies 
pydantic
python-dotenv
wikipedia
faiss-cpu